Divide windows
+instructions
