+++ CoursesManager/WpfApp1/PersonalAccountWindow.xaml
@@ -1,0 +1,60 @@
+﻿<Window x:Class="WpfApp1.PersonalAccountWindow"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:WpfApp1"
+        mc:Ignorable="d"
+        xmlns:CoursesManagerLib="clr-namespace:CoursesManagerLib;assembly=CoursesManagerLib"
+        Title="PersonalAccountWindow" Height="800" Width="1200" ResizeMode="NoResize">
+    <Grid x:Name="GridPersonalClient"   Background="#FFD67979" Height="780" Width="1190" >
+        <ComboBox x:Name="BoxPersFormat" HorizontalAlignment="Left" Height="45" Margin="106,60,0,0" VerticalAlignment="Top" Width="180" ScrollViewer.VerticalScrollBarVisibility="Visible" FontSize="20" SelectedIndex="0" SelectionChanged="BoxPersFormat_SelectionChanged">
+            <CoursesManagerLib:Format>Individual</CoursesManagerLib:Format>
+            <CoursesManagerLib:Format>Group</CoursesManagerLib:Format>
+        </ComboBox>
+        <ComboBox x:Name="BoxPersLanguage" HorizontalAlignment="Left" Height="45" Margin="106,140,0,0" VerticalAlignment="Top" Width="180" FontSize="20" SelectedIndex="0">
+            <TextBlock>English</TextBlock>
+            <TextBlock>Spanish</TextBlock>
+            <TextBlock>French</TextBlock>
+            <TextBlock>German</TextBlock>
+        </ComboBox>
+        <ComboBox x:Name="BoxPersLevel" HorizontalAlignment="Left" Height="45" Margin="106,220,0,0" VerticalAlignment="Top" Width="180" FontSize="20"  SelectedIndex="0">
+            <TextBlock>Primary</TextBlock>
+            <TextBlock>Medium</TextBlock>
+            <TextBlock>Profy</TextBlock>
+        </ComboBox>
+        <ComboBox x:Name="BoxPersIntensity" HorizontalAlignment="Left" Height="45" Margin="106,300,0,0" VerticalAlignment="Top" Width="180" FontSize="20" SelectedIndex="0" SelectionChanged="BoxPersIntensity_SelectionChanged">
+            <TextBlock>1</TextBlock>
+            <TextBlock>2</TextBlock>
+            <TextBlock>3</TextBlock>
+        </ComboBox>
+        <Label x:Name="LabelPersFormat" Content="Format" HorizontalAlignment="Left" Height="45" Margin="350,60,0,0" VerticalAlignment="Top" Width="400" Background="#FFE7D7D7" FontSize="20"/>
+        <Label x:Name="LabelPersLanguage" Content="Language" HorizontalAlignment="Left" Height="45" Margin="350,140,0,0" VerticalAlignment="Top" Width="400" Background="#FFE7D7D7" FontSize="20"/>
+        <Label x:Name="LabelPersLevel" Content="Level" HorizontalAlignment="Left" Height="45" Margin="350,220,0,0" VerticalAlignment="Top" Width="400" Background="#FFE7D7D7" FontSize="20"/>
+        <Label x:Name="LabelPersIntensity" Content="Intensity" HorizontalAlignment="Left" Height="43" Margin="350,300,0,0" VerticalAlignment="Top" Width="400" Background="#FFE7D7D7" FontSize="20"/>
+        <Label x:Name="LabelPersDuration" HorizontalAlignment="Left" Height="45" Margin="350,380,0,0" VerticalAlignment="Top" Width="400" Background="#FFE7D7D7" FontSize="20" Content="Duration (months)"/>
+        <Label x:Name="LabelPersPrice" HorizontalAlignment="Left" Height="45" Margin="350,460,0,0" VerticalAlignment="Top" Width="400" Background="#FFE7D7D7" FontSize="20" Content="Price (month)"/>
+        <Label x:Name="LabelPersDurationAvto" HorizontalAlignment="Left" Height="45" Margin="106,380,0,0" VerticalAlignment="Top" Width="180" Background="#FFE7D7D7" FontSize="20" Content="10"/>
+        <Label x:Name="LabelPersPriceAvto" HorizontalAlignment="Left" Height="45" Margin="106,460,0,0" VerticalAlignment="Top" Width="180" Background="#FFE7D7D7" FontSize="20" Content="3200"/>
+        <Label x:Name="LabelPersID" HorizontalAlignment="Left" Height="45" Margin="350,540,0,0" VerticalAlignment="Top" Width="400" Background="#FFE7D7D7" FontSize="20" Content="ID"/>
+        <Button x:Name="ButtonPersAddClaim" Content="Add Your Claim" HorizontalAlignment="Left" Height="84" Margin="106,610,0,0" VerticalAlignment="Top" Width="219" FontSize="20" Click="ButtonPersAddClaim_Click" Background="#FFC03956"/>
+        <TextBox x:Name="TextBoxPersID"  IsReadOnly="True" HorizontalAlignment="Left" Height="45" Margin="106,540,0,0" TextWrapping="Wrap" VerticalAlignment="Top" Width="180" Background="#FFE7D7D7" FontSize="20"/>
+        <TextBox x:Name="TextBoxPersName" IsReadOnly="True" HorizontalAlignment="Left" Height="45" Margin="868,140,0,0" TextWrapping="Wrap" VerticalAlignment="Top" Width="205" Background="#FFE7D7D7" FontSize="20"/>
+        <TextBox x:Name="TextBoxPersSurname" IsReadOnly="True" HorizontalAlignment="Left" Height="45" Margin="868,300,0,0" TextWrapping="Wrap" VerticalAlignment="Top" Width="205" Background="#FFE7D7D7" FontSize="20"/>
+        <Label x:Name="LabelPersSurname" HorizontalAlignment="Left" Height="45" Margin="868,220,0,0" VerticalAlignment="Top" Width="205" Background="#FFE7D7D7" FontSize="20" Content="Surname"/>
+        <Label x:Name="LabelPersName" HorizontalAlignment="Left" Height="45" Margin="868,60,0,0" VerticalAlignment="Top" Width="205" Background="#FFE7D7D7" FontSize="20" Content="Name"/>
+        <Button x:Name="ButtonPersClientWindow" Content="Return" HorizontalAlignment="Left" Height="84" Margin="854,610,0,0" VerticalAlignment="Top" Width="219" IsCancel="True" FontSize="20" Background="#FF6D8EDA" Click="ButtonPersClientWindow_Click"/>
+        <TextBox x:Name="TextBoxMoney" HorizontalAlignment="Left" Height="45" Margin="868,380,0,0" TextWrapping="Wrap"  VerticalAlignment="Top" Width="205" Background="#FFE7D7D7" FontSize="20"/>
+        <Button x:Name="ButtonChangeMoney" Content="Change count of money" HorizontalAlignment="Left" Margin="868,460,0,0" VerticalAlignment="Top" Width="205" Height="45" FontSize="17" Click="ButtonChangeMoney_Click"/>
+        <TextBox x:Name="textBoxGetMoney" HorizontalAlignment="Left" Height="45" Margin="868,540,0,0" TextWrapping="Wrap" IsReadOnly="True" VerticalAlignment="Top" Width="205" Background="#FFE7D7D7" FontSize="20"/>
+        <ToolBar x:Name="PerstoolBar" Background="#FFF67979" HorizontalAlignment="Left" Height="55" VerticalAlignment="Top" Width="1190">
+            <Menu Height="40" Margin="860,5,0,-5.4" VerticalAlignment="Top" Width="200">
+                <MenuItem Header="Instruction" FontSize="22">
+                    <MenuItem x:Name="EngPersInstruction" Header="English" Click="EngPersInstruction_Click" />
+                    <MenuItem x:Name="RusPersInstruction" Header="Russian" Click="RusPersInstruction_Click"  />
+                    <Separator />
+                </MenuItem>
+            </Menu>
+        </ToolBar>
+    </Grid>
+</Window>
