+++ CoursesManager/WpfApp1/bin/Debug/PersRusInstruction.txt
@@ -1,0 +1,13 @@
+Add yuor claim - позволяет добавить вашу заявку, для этого нужно заполнить следующие поля:
+  Name - имя
+  Surname - фамилия
+  Format - формат занятий
+  Language - язык
+  Level - уровень подготовки языка
+  Intensity - интенсивность занятий в неделю
+Поля,заполняющиеся автоматически:
+  Duration - длительность курса (зависит от интенсивности)
+  Price - цена (зависит от формата и интенсивности)
+  ID - ваш персональный номер 
+Change count of money - количество денег на вашем счету
+Return - вернуться в окно гостя
