+++ CoursesManager/WpfApp1/bin/Debug/RusAdminInstruction.txt
@@ -1,0 +1,8 @@
+Load - загрузить информацию из файла
+Save - сохранить информацию в файл
+SetAttendance - отметить посещаемость 
+GetAttendance - посмотреть посещаемость
+Show Groups/Clients/Claims - показать список Групп/Клиентов/Заявок (после нажатия кнопки Admission )
+Admission - зачисляет клиентов, составляет список заявок
+ShowGroupsID - позволяет посмотреть список групп для определенного клиента по его ID 
+ClientWindow  - переходит в гостевое окно клиента
