+++ CoursesManager/WpfApp1/bin/Debug/RusInstruction.txt
@@ -1,0 +1,14 @@
+Add yuor claim - позволяет добавить вашу заявку, для этого нужно заполнить следующие поля:
+  Name - ваше имя
+  Surname - ваша фамилия
+  Format - формат занятий
+  Language - язык
+  Level - уровень подготовки языка
+  Intensity - интенсивность занятий в неделю
+Поля,заполняющиеся автоматически:
+  Duration - длительность курса (зависит от интенсивности)
+  Price - цена (зависит от формата и интенсивности)
+  ID - ваш персональный номер,с его помощью вы можете зайти в свой личный кабинет 
+Account - введите ваш ID, чтобы войти в личный кабинет
+Admin Window - вернуться в окно администратора
+
